OS2UOL-72: Robots improvements taken a step further
