+++ config/productionDomains.ts
@@ -1,0 +1,24 @@
+const prodDomains = [
+  'aabenaalborg.dk',
+  'www.aabenaalborg.dk',
+  'ulfiaarhus.dk',
+  'www.ulfiaarhus.dk',
+  'udsynmodarbejdsliv.dk',
+  'www.udsynmodarbejdsliv.dk',
+  'mitvadehav.dk',
+  'www.mitvadehav.dk',
+  'ude.nu',
+  'www.ude.nu',
+  'udoglaer.randers.dk',
+  'www.udoglaer.randers.dk',
+  'udoglaer.dk',
+  'www.udoglaer.dk',
+  'laeringsportalenskive.dk',
+  'www.laeringsportalenskive.dk',
+  'rum.thisted.dk',
+  'www.rum.thisted.dk',
+  'udoglaer.vejle.dk',
+  'www.udoglaer.vejle.dk',
+  'klcviborg.dk',
+  'www.klcviborg.dk',
+];
