+++ server/middleware/robots.ts
@@ -1,3 +1,5 @@
+import { productionDomains } from '~/config/productionDomains';
+
 export default defineEventHandler((event) => {
   if (event.node.req.url !== '/robots.txt') {
     return;
@@ -25,32 +27,7 @@ export default defineEventHandler((event) => {
     Disallow: /
     `;
 
-  const prodDomains = [
-    'aabenaalborg.dk',
-    'www.aabenaalborg.dk',
-    'ulfiaarhus.dk',
-    'www.ulfiaarhus.dk',
-    'udsynmodarbejdsliv.dk',
-    'www.udsynmodarbejdsliv.dk',
-    'mitvadehav.dk',
-    'www.mitvadehav.dk',
-    'ude.nu',
-    'www.ude.nu',
-    'udoglaer.randers.dk',
-    'www.udoglaer.randers.dk',
-    'udoglaer.dk',
-    'www.udoglaer.dk',
-    'laeringsportalenskive.dk',
-    'www.laeringsportalenskive.dk',
-    'rum.thisted.dk',
-    'www.rum.thisted.dk',
-    'udoglaer.vejle.dk',
-    'www.udoglaer.vejle.dk',
-    'klcviborg.dk',
-    'www.klcviborg.dk',
-  ];
-
-  if (host && prodDomains.includes(host)) {
+  if (host && productionDomains.includes(host)) {
     robotsTxt = prodRobotTxt;
   } else {
     robotsTxt = stagingRobotTxt;
